refactor: 정적 리소스 요청을 처리하는 컨트롤러 추가
