+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,15 +1,13 @@
 package org.apache.coyote.http11;
 
-import java.io.File;
 import java.io.IOException;
 import java.net.Socket;
-import java.net.URL;
-import java.util.Optional;
 import nextstep.jwp.controller.HomeController;
 import nextstep.jwp.controller.LoginController;
 import nextstep.jwp.controller.RegisterController;
 import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
+import nextstep.jwp.controller.StaticResourceController;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -46,49 +44,28 @@ public class Http11Processor implements Runnable, Processor {
     }
 
     private HttpResponse createResponse(final HttpRequest request) {
-        HttpResponse response = createResponseFromRequest(request);
+        final HttpResponse response = findController(request).doService(request);
+
         addContentTypeToResponse(request, response);
         addCookieForSessionToResponse(request, response);
+
         return response;
     }
 
-    private HttpResponse createResponseFromRequest(final HttpRequest request) {
-        Optional<Controller> controllerOptional = findController(request);
-
-        if (controllerOptional.isPresent()) {
-            Controller controller = controllerOptional.get();
-            return controller.doService(request);
-        }
-
-        return createStaticResourceResponse(request);
-    }
-
-    private Optional<Controller> findController(final HttpRequest request) {
+    private Controller findController(final HttpRequest request) {
         if (request.getUriPath().equals("/")) {
-            return Optional.of(new HomeController());
+            return new HomeController();
         }
 
         if (request.getUriPath().equals("/login")) {
-            return Optional.of(new LoginController());
+            return new LoginController();
         }
 
         if (request.getUriPath().equals("/register")) {
-            return Optional.of(new RegisterController());
+            return new RegisterController();
         }
 
-        return Optional.empty();
-    }
-
-    private HttpResponse createStaticResourceResponse(final HttpRequest request) {
-        if (hasStaticResourceFile(request.getUriPath())) {
-            return HttpResponse.ok().fileBody(request.getUriPath()).build();
-        }
-        return HttpResponse.notFound().build();
-    }
-
-    private boolean hasStaticResourceFile(final String uri) {
-        final URL resource = getClass().getClassLoader().getResource("static" + uri);
-        return resource != null && new File(resource.getFile()).isFile();
+        return new StaticResourceController();
     }
 
     private void addContentTypeToResponse(final HttpRequest request, final HttpResponse response) {
